Refactored Code and Addressed comments
